+++ CHANGELOG.md
@@ -18,6 +18,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Security
 
+## v7.7.0 - 2020-07-01
+
+### Changed
+
+- update @chevre/factory\
+- webAPIインターフェースをChevreへ完全移行
+
 ## v7.6.2 - 2020-06-29
 
 ### Changed